NINGG-10048 - Added visa instalment saved card flow

+++ payment-sdk-core/src/main/java/payment/sdk/android/core/Order.kt
@@ -19,7 +19,7 @@ class Order {
     // Other classes
     class Amount {
         var currencyCode: String? = null
-        var value: Int? = 0
+        var value: Double? = 0.0
     }
 
     class Links {
